Fix scrollbar behavior
