+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,10 +1,83 @@
-.ingredients{
+.ingredients {
     display: flex;
     flex-direction: column;
-    padding: 40px 0;
-    margin-left: 20px;   
+    margin-right: 20px;
     width: var(--width-column-main);
     box-sizing: border-box;
+}
 
-    border: 2px solid var(--color-blue);
+.menu {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    padding-top: 40px;
+    background-color: var(--background-color);
+    height: var(--height-menu);
+    box-sizing: border-box;
+}
+
+.navigation {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.navLink {
+    display: flex;
+    position: relative;
+    width: calc(var(--width-column-main) / 3);
+    align-content: center;
+    justify-content: center;
+    color: var(--text-secondary-color);
+    transition: all 0.2s;
+    max-height: 56px;
+    box-sizing: border-box;
+}
+
+.navLink:hover {
+    cursor: hover;
+    color: var(--text-primary-color);
+}
+
+.navLinkActive {
+    composes: navLink;
+    border-bottom: 2px solid var(--color-blue);
+    color: var(--text-primary-color);
+}
+
+.sectionFood {
+    position: relative;
+    display: flex;
+    flex-direction: column;
+    margin: 40px 0 0;
+}
+
+.cards {
+    display: grid;
+    grid-template-rows: auto 1fr;
+    grid-template-columns: 50% 50%;
+}
+
+.card {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    align-items: center;
+    margin: 16px;
+}
+
+.cardPrice {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+
+.cardTitle {
+    text-align: center;
+}
+
+.count {
+    position: absolute;
+    top: 0;
+    right: 0;
 }
