+++ src/components/scrollbox/scrollbox.module.css
@@ -10,8 +10,8 @@
 .scrollBarContainer {
     position: absolute;
     right: 0px;
-    bottom: 5px;
-    top: 5px;
+    bottom: 0px;
+    top: 0px;
     width: 8px;
     background-color: var(--background-light-color);
 }
